requirements.txt
matplotlib==3.6.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.3.0
plotly==5.9.0
streamlit==1.15.2
